server/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.1
psycopg2==2.9.10
python-dotenv==1.0.0
Flask-Bcrypt==1.0.1
Flask-Session==0.5.0

